+++ src/lib/validators/iban-validator.ts
@@ -1,5 +1,6 @@
-import { AbstractControl, ValidationErrors } from '@angular/forms';
+import { AbstractControl, NG_VALIDATORS, ValidationErrors, Validator } from '@angular/forms';
 import * as IBAN from 'iban';
+import { Directive, Input } from '@angular/core';
 
 /**
  * IBAN validation for reactive FormControls
@@ -10,4 +11,17 @@ export function ibanValidator(control: AbstractControl): ValidationErrors {
     }
 
     return null;
+}
+
+@Directive({
+    selector: 'input[iban][ngModel]',
+    providers: [{ provide: NG_VALIDATORS, useExisting: IbanValidatorDirective, multi: true }]
+})
+export class IbanValidatorDirective implements Validator {
+    @Input('iban')
+    public shouldValidate: boolean;
+
+    public validate(control: AbstractControl): ValidationErrors | null {
+        return this.shouldValidate ? ibanValidator(control) : null;
+    }
 }